+++ src/iop/_business_operation.py
@@ -47,6 +47,7 @@ class _BusinessOperation(_BusinessHost):
 
     def _dispatch_on_init(self, host_object: Any) -> None:
         """For internal use only."""
+        self._debugpy(host_object=host_object)
         create_dispatch(self)
         self.on_init()
         return