+++ src/iop/_business_process.py
@@ -141,6 +141,7 @@ class _BusinessProcess(_BusinessHost):
     def _dispatch_on_init(self, host_object: Any) -> None:
         """For internal use only."""
         self._restore_persistent_properties(host_object)
+        self._debugpy(host_object=host_object)
         create_dispatch(self)
         self.on_init()
         self._save_persistent_properties(host_object)