+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add namespace selection in the `iop` command
   - `-n` or `--namespace` option to select a namespace
+- Initial remote debugging support
 
 ### Changed
 - Change how `iris` module is loaded